refactor(ipc): address review feedback
- Move the reserved control topics and is_control_topic into a dedicated `control`
  module instead of living in reachability.
- Extract a `prune` helper for the registry cleanup duplicated in track() and Drop.
- Rewrite the ReachabilityTracker public doc from a consumer's perspective.
- Merge the two cfg_attr attributes on Reachability into one.
- Move the WASM ReachabilityTracker/Handle wrappers into their own wasm/reachability.rs.
