+++ crates/bitwarden-ipc/src/wasm/mod.rs
@@ -4,9 +4,11 @@ mod generic_session_repository;
 mod ipc_client;
 mod js_session_repository;
 mod message;
+mod reachability;
 
 // Re-export types to make sure wasm_bindgen picks them up
 pub use communication_backend::*;
 pub use discover::*;
 pub use ipc_client::*;
 pub use js_session_repository::*;
+pub use reachability::*;
